cleaned up read and write functions

+++ src/disk.h
@@ -42,16 +42,16 @@ int bmfs_disk_delete_file(struct BMFSDisk *disk, const char *filename);
 int bmfs_disk_format(struct BMFSDisk *disk);
 
 int bmfs_initialize(char *diskname, char *size, char *mbr, char *boot, char *kernel);
-unsigned long long bmfs_read(struct BMFSDisk *disk,
-                             const char *filename,
-                             void * buf,
-                             unsigned long long len,
-                             unsigned long long off);
+int bmfs_read(struct BMFSDisk *disk,
+              const char *filename,
+              void * buf,
+              uint64_t len,
+              uint64_t off);
 int bmfs_write(struct BMFSDisk *disk,
                const char *filename,
                const void *buf,
-               size_t len,
-               off_t off);
+               uint64_t len,
+               uint64_t off);
 void bmfs_readfile(struct BMFSDisk *disk, const char *filename);
 void bmfs_writefile(struct BMFSDisk *disk, const char *filename);
